refactor: frame 패키지 추가
