+++ src/main/java/bowling/frame/Frames.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.frame;
 
 import java.util.ArrayList;
 import java.util.List;
